[REFAC]: publisher 패키지 파일 분리 + facade 정의 (이슈 #386)
#385 (Publisher 통합 모듈화 메타) Sub 1 — 외부 facade 정의 + 파일 분리로 다른
sub 들의 prerequisite 완성.

파일 분리 (3 파일):
- publisher.go: Publisher struct + 생성자 + 공통 Kafka helpers (buildMessage /
  crawlTopic / newJobID) + PriorityResolver interface + 패키지 doc 갱신
- chain.go (신규): PublishChained 메소드 (구 Publish) + normalizeURLs /
  acquireViaGuard / acquireIngestion helpers
- guard.go (신규): IngestionLock / PipelineGuard interface + atomic wrapper +
  SetGate / SetNormalizer / SetIngestionLock / SetPipelineGuard

메소드 rename:
- Publisher.Publish → Publisher.PublishChained
- general.JobPublisher.Publish → PublishChained (인터페이스 정의)
- search.JobPublisher.Publish → PublishChained (인터페이스 정의)
- caller 호출처 (parser_worker, search handler) 동기 갱신
- 테스트 mock recordingPublisher.Publish → PublishChained
- 테스트 호출 pub.Publish → pub.PublishChained

향후 Sub 메소드 (PublishSeed/PublishRetry/PublishUpgrade) 는 각 Sub 가 추가 —
본 Sub 는 chain 분리 + facade 골격만 완성.

검증:
- go build + go vet 통과
- 전체 38 패키지 race test 통과
