+++ internal/processor/fetcher/domain/general/types.go
@@ -36,9 +36,10 @@ type Fetcher interface {
 }
 
 // JobPublisher 는 카테고리/목록 페이지에서 발견된 URL 을 다음 CrawlJob 으로 연결하는 인터페이스입니다.
-// publisher.Publisher 가 본 인터페이스를 구현하며, ChainHandler 에 주입됩니다.
+// publisher.Publisher 가 본 인터페이스를 구현하며 (PublishChained 메소드 — 이슈 #386),
+// ChainHandler 에 주입됩니다.
 //
 // JobPublisher dispatches CrawlJobs discovered from list/category pages.
 type JobPublisher interface {
-	Publish(ctx context.Context, crawlerName string, urls []string, targetType core.TargetType, timeout time.Duration) error
+	PublishChained(ctx context.Context, crawlerName string, urls []string, targetType core.TargetType, timeout time.Duration) error
 }
